+++ source/nxcore/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_library(nxcore INTERFACE "./primitives.h" "./helpers.h")
 
-add_library(NinetailsXEngine SHARED "./engine/game.h" "./engine/game.cpp")
+add_library(NinetailsXEngine SHARED "./engine/engine.h" "./engine/engine.cpp")
 target_link_libraries(NinetailsXEngine PUBLIC nxcore)
