+++ source/nxcore/engine.h
